# backend/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tela REST API",
    "version": "2026-09-28",
    "summary": "HTTP API of tela, a self-hostable markdown-native team wiki.",
...
    "contact": {
      "name": "tela",
      "url": "https://telawiki.com/support",
      "email": "tela@telawiki.com"
    },
    "license": {
      "name": "AGPL-3.0-or-later",
      "url": "https://github.com/zcag/tela/blob/main/LICENSE"
    }
  },
...
  "servers": [
    {
      "url": "https://telawiki.com",
      "description": "The instance serving this document. A self-hosted instance rewrites this to its own origin."
    }
  ],
  "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }],
...
  "paths": {
...
    "/api/pages/{id}/md": {
      "parameters": [{ "$ref": "#/components/parameters/PathId" }],
      "get": {
        "tags": ["pages"],
        "operationId": "exportPageMarkdown",
        "summary": "Download a page as markdown",
        "description": "Canonical frontmatter plus the body — the same rendering the zip export and the WebDAV sync surface produce.",
        "responses": {
          "200": {
            "description": "Markdown.",
            "headers": {
              "Content-Disposition": { "description": "`attachment; filename=\"<page-slug>.md\"`", "schema": { "type": "string" } }
            },
            "content": { "text/markdown": { "schema": { "type": "string" } } }
          },
          "400": { "$ref": "#/components/responses/InvalidId" },
          "403": { "$ref": "#/components/responses/ForbiddenOrMissingPage" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
...
    "/api/public/spaces/{id}/feed.xml": {
      "parameters": [{ "$ref": "#/components/parameters/PathId" }],
      "get": {
        "tags": ["public"],
        "operationId": "getPublicSpaceFeed",
        "summary": "RSS feed of a published space",
        "description": "The 50 most recent top-level pages, newest first.",
        "security": [],
        "responses": {
          "200": {
            "description": "RSS 2.0.",
            "content": { "application/rss+xml": { "schema": { "type": "string" } } }
          },
          "400": { "$ref": "#/components/responses/InvalidId" },
          "404": { "$ref": "#/components/responses/NoSuchPublicSpace" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
...
    "/api/public/by-handle/{handle}": {
      "parameters": [{ "$ref": "#/components/parameters/Handle" }],
      "get": {
        "tags": ["public"],
        "operationId": "getPublicByHandle",
        "summary": "A public account home",
        "description": "Resolves one namespace of users and organizations (a user wins a collision) and returns that account's published spaces and latest posts. An account with nothing published is a 404, same as an unknown handle.",
        "security": [],
        "responses": {
          "200": {
            "description": "The account home.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HandleHome" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A personal access token: `Authorization: Bearer tela_pat_<43 chars>`. Scope `read` allows GET and HEAD only; `write` and `admin` allow any method. A token may be pinned to a single space."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tela_session",
        "description": "The browser session cookie set by `POST /api/auth/login`. HttpOnly, SameSite=Lax, 30 days."
      }
    },
    "parameters": {
      "PathId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64", "minimum": 1 }
      },
...
      "Handle": {
        "name": "handle",
        "in": "path",
        "required": true,
        "description": "A username or an organization slug — one namespace, case-insensitive.",
        "schema": { "type": "string" }
      }
    },
    "responses": {
...
      "InvalidId": {
        "description": "`invalid_id` — a path parameter was not a positive integer.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "ForbiddenOrMissingPage": {
        "description": "`forbidden` — **also returned when the page does not exist**, so page ids cannot be enumerated. When the page happens to live in a published space, `code` is `forbidden_public` and `public_path` points at the no-login reader. Also `api_key_scope` / `api_key_space_scope`.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "`not_found`.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NoSuchPublicSpace": {
        "description": "`not_found` — no such published space. Private and non-existent are indistinguishable.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "Internal": {
        "description": "`internal` — server error.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "The error envelope for every JSON route. Branch on `code`, not on the message.",
        "required": ["error", "code"],
        "properties": {
          "error": { "type": "string", "description": "Human-readable message. Not stable — do not parse it." },
          "code": { "type": "string", "description": "Machine-readable code.", "examples": ["forbidden", "not_found", "api_key_scope"] },
          "public_path": { "type": "string", "description": "Only with `code: forbidden_public` — where to read this page without an account." }
        }
      },
      "Timestamp": {
        "type": "string",
        "description": "`YYYY-MM-DD HH:MM:SS` in UTC.",
        "examples": ["2026-09-09 14:03:21"]
      },
...
      "CardMeta": {
        "type": "object",
        "description": "Index-card metadata computed from a page, so a listing can be rendered without fetching bodies.",
        "properties": {
          "kind": { "type": "string", "enum": ["deck", "sheet"], "description": "Absent for an ordinary prose page." },
          "excerpt": { "type": "string" },
          "reading_minutes": { "type": "integer" },
          "cover": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "HandleHome": {
        "type": "object",
        "required": ["kind", "handle", "name", "bio", "spaces", "posts"],
        "properties": {
          "kind": { "type": "string", "enum": ["user", "org"] },
          "handle": { "type": "string" },
          "name": { "type": "string" },
          "bio": { "type": "string", "description": "Empty for organizations." },
          "spaces": {
            "type": "array",
            "description": "Published spaces, most recently active first.",
            "items": {
              "type": "object",
              "required": ["id", "name", "slug", "description", "page_count", "updated_at"],
              "properties": {
                "id": { "type": "integer", "format": "int64" },
                "name": { "type": "string" },
                "slug": { "type": "string" },
                "description": { "type": "string" },
                "page_count": { "type": "integer", "format": "int64" },
                "updated_at": { "type": "string" }
              }
            }
          },
          "posts": {
            "type": "array",
            "description": "Up to 6 recent top-level public pages across those spaces.",
            "items": {
              "allOf": [
                {
                  "type": "object",
                  "required": ["space_id", "space_name", "space_slug", "id", "title", "created_at", "updated_at"],
                  "properties": {
                    "space_id": { "type": "integer", "format": "int64" },
                    "space_name": { "type": "string" },
                    "space_slug": { "type": "string" },
                    "id": { "type": "integer", "format": "int64" },
                    "title": { "type": "string" },
                    "created_at": { "$ref": "#/components/schemas/Timestamp" },
                    "updated_at": { "$ref": "#/components/schemas/Timestamp" }
                  }
                },
                { "$ref": "#/components/schemas/CardMeta" }
              ]
            }
          }
        }
      },
...
    }
  }
}